+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 FEATURES:
 * **New Datasource**: `hcloud_server` Lookup a server.
 
+IMPROVEMENTS:
+* Add API token length validation
+
 ## 1.7.0 (December 18, 2018)
 
 FEATURES: